requirements.txt
# Core data science libraries
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Statistical analysis
statsmodels>=0.13.0
pingouin>=0.5.0
lifelines>=0.27.0

# Machine Learning
xgboost>=1.5.0
lightgbm>=3.3.0
torch>=2.0.0
torchvision>=0.15.0

# GUI
tkinter
customtkinter>=5.0.0

# System utilities
psutil>=5.8.0
memory-profiler>=0.60.0
tqdm>=4.62.0

# Visualization backends
kaleido>=0.2.1

# Data processing
openpyxl>=3.0.9
xlrd>=2.0.1

# AI and Natural Language Processing
openai>=1.0.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-google-genai>=0.0.5

# Image processing and OCR
pillow>=9.0.0
opencv-python>=4.5.0
pytesseract>=0.3.10
easyocr>=1.6.0

# MCP (Model Context Protocol) support
httpx>=0.24.0
websockets>=11.0.0
pydantic>=2.0.0

# Advanced data analysis
spacy>=3.4.0
transformers>=4.20.0

# File format support
python-docx>=0.8.11
pdfplumber>=0.7.0

# API utilities
requests>=2.28.0
aiohttp>=3.8.0

# Configuration management
python-dotenv>=0.19.0
pyyaml>=6.0

# Professional Features - 추가 필수 라이브러리
jinja2>=3.1.0           # プロフェッショナルレポート生成
cryptography>=3.4.8

# 高度統計解析・最適化
statsmodels>=0.14.0
pingouin>=0.5.3
lifelines>=0.27.0
numba>=0.58.0
psutil>=5.9.0
scikit-learn>=1.3.0
scipy>=1.11.0

# 並列処理・最適化
joblib>=1.3.0
tqdm>=4.66.0

# オプション: ベイズ統計（高度機能）
# pymc>=5.7.0
# arviz>=0.16.0

# オプション: 分散処理（大規模データ）
# dask[complete]>=2023.8.0     # セキュリティ機能

# 統計解析専門ライブラリ
statsmodels>=0.14.0
pingouin>=0.5.3
scikit-learn>=1.3.0
scikit-image>=0.21.0
arviz>=0.15.0           # ベイズ診断

# 高度な機械学習
catboost>=1.2.0
optuna>=3.1.0           # ハイパーパラメータ最適化
imbalanced-learn>=0.10.0

# 深層学習フレームワーク
tensorflow>=2.13.0
torchaudio>=2.0.0
keras-tuner>=1.3.0

# 時系列解析
arch>=5.3.0             # GARCH モデル
statsforecast>=1.4.0
tsfresh>=0.20.0         # 時系列特徴抽出
prophet>=1.1.0          # Facebook Prophet

# 可視化・グラフィックス
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
bokeh>=3.1.0
altair>=5.0.0
dash>=2.10.0            # インタラクティブダッシュボード

# GUI フレームワーク
PyQt5>=5.15.0
streamlit>=1.28.0       # Webベース GUI - Mac M2対応
plotly>=5.17.0          # インタラクティブプロット

# データ処理・I/O
xlsxwriter>=3.1.0
pyarrow>=12.0.0         # Parquet サポート
h5py>=3.8.0
joblib>=1.2.0

# プロフェッショナル機能
fpdf2>=2.7.0            # PDF レポート生成
jinja2>=3.1.0           # レポートテンプレート
networkx>=3.1.0         # ネットワーク解析
umap-learn>=0.5.3       # UMAP 次元削減
hdbscan>=0.8.29         # HDBSCAN クラスタリング

# ユーティリティ
tqdm>=4.65.0
psutil>=5.9.0           # システム監視
memory-profiler>=0.60.0 # メモリプロファイリング 